backend/foodgram/requirements.txt
django==2.2.19
django-filter==21.1
djangorestframework==3.12.4
PyJWT==2.1.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3
requests==2.26.0
gunicorn==20.0.4
sqlparse==0.3.1
pytz==2020.1
djoser==2.1.0
drf-extra-fields==3.4.0